Remove VariantDatasetFunctions.
Move remaining methods to method objects.
