+++ src/main/scala/is/hail/variant/MatrixTable.scala
@@ -258,6 +258,28 @@ object MatrixTable {
     val (first, others) = (datasets.head, datasets.tail)
     first.copyLegacy(rdd = first.sparkContext.union(datasets.map(_.rdd)))
   }
+
+  def fromKeyTable(kt: Table): MatrixTable = {
+    val vType: Type = kt.keyFields.map(_.typ) match {
+      case Array(t@TVariant(_, _)) => t
+      case arr => fatal("Require one key column of type Variant to produce a variant dataset, " +
+        s"but found [ ${ arr.mkString(", ") } ]")
+    }
+
+    val rdd = kt.keyedRDD()
+      .map { case (k, v) => (k.asInstanceOf[Row].get(0), v) }
+      .filter(_._1 != null)
+      .mapValues(a => (a: Annotation, Iterable.empty[Annotation]))
+
+    val metadata = VSMMetadata(
+      saSignature = TStruct.empty(),
+      vSignature = vType,
+      vaSignature = kt.valueSignature,
+      globalSignature = TStruct.empty())
+
+    MatrixTable.fromLegacy(kt.hc, metadata,
+      VSMLocalValue(Annotation.empty, Array.empty[Annotation], Array.empty[Annotation]), rdd)
+  }
 }
 
 case class VSMSubgen(
@@ -2656,5 +2678,52 @@ class MatrixTable(val hc: HailContext, val metadata: VSMMetadata,
     info(s"Wrote all $blockCount blocks of $nRows x $nCols matrix with block size $blockSize.")
     
     hadoop.writeTextFile(dirname + "/_SUCCESS")(out => ())
+  }
+
+  // FIXME remove when filter_alleles tests are migrated to Python
+  def filterAlleles(filterExpr: String, variantExpr: String = "",
+    keep: Boolean = true, subset: Boolean = true, leftAligned: Boolean = false, keepStar: Boolean = false): MatrixTable = {
+    if (!genotypeSignature.isOfType(Genotype.htsGenotypeType))
+      fatal(s"filter_alleles: genotype_schema must be the HTS genotype schema, found: ${ genotypeSignature }")
+
+    val genotypeExpr = if (subset) {
+      """
+g = let newpl = if (isDefined(g.PL))
+        let unnorm = range(newV.nGenotypes).map(newi =>
+            let oldi = gtIndex(newToOld[gtj(newi)], newToOld[gtk(newi)])
+             in g.PL[oldi]) and
+            minpl = unnorm.min()
+         in unnorm - minpl
+      else
+        NA: Array[Int] and
+    newgt = gtFromPL(newpl) and
+    newad = if (isDefined(g.AD))
+        range(newV.nAlleles).map(newi => g.AD[newToOld[newi]])
+      else
+        NA: Array[Int] and
+    newgq = gqFromPL(newpl) and
+    newdp = g.DP
+ in { GT: Call(newgt), AD: newad, DP: newdp, GQ: newgq, PL: newpl }
+        """
+    } else {
+      // downcode
+      s"""
+g = let newgt = gtIndex(oldToNew[gtj(g.GT)], oldToNew[gtk(g.GT)]) and
+    newad = if (isDefined(g.AD))
+        range(newV.nAlleles).map(i => range(v.nAlleles).filter(j => oldToNew[j] == i).map(j => g.AD[j]).sum())
+      else
+        NA: Array[Int] and
+    newdp = g.DP and
+    newpl = if (isDefined(g.PL))
+        range(newV.nGenotypes).map(gi => range(v.nGenotypes).filter(gj => gtIndex(oldToNew[gtj(gj)], oldToNew[gtk(gj)]) == gi).map(gj => g.PL[gj]).min())
+      else
+        NA: Array[Int] and
+    newgq = gqFromPL(newpl)
+ in { GT: Call(newgt), AD: newad, DP: newdp, GQ: newgq, PL: newpl }
+        """
+    }
+
+    FilterAlleles(this, filterExpr, variantExpr, genotypeExpr,
+      keep = keep, leftAligned = leftAligned, keepStar = keepStar)
   }
 }
