+++ src/main/scala/is/hail/variant/package.scala
@@ -23,6 +23,4 @@ package object variant {
   }
 
   implicit def toRichIterableGenotype(ig: Iterable[Annotation]): RichIterableGenotype = new RichIterableGenotype(ig)
-
-  implicit def toVDSFunctions(vds: MatrixTable): VariantDatasetFunctions = new VariantDatasetFunctions(vds)
 }
