+++ src/test/scala/is/hail/io/HardCallsSuite.scala
@@ -9,7 +9,7 @@ import org.testng.annotations.Test
 class HardCallsSuite extends SparkSuite {
   @Test def test() {
     val p = forAll(MatrixTable.gen(hc, VSMSubgen.random)) { vds =>
-      val hard = vds.hardCalls()
+      val hard = vds.annotateGenotypesExpr("g = {GT: g.GT}")
       assert(hard.queryGenotypes("gs.map(g => g.GT).counter()") ==
         vds.queryGenotypes("gs.map(g => g.GT).counter()"))
